codec.py
